+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.2.9] - 2024-04-18
+## [1.2.9] - 2024-04-19
 
 ### Added